Add ROSMessages and ROSTransforms
ROS messages can reference transforms from /tf and /tf_static

We provided access to transforms in getMetadata() in converters,
but not in convertMessage(). With this change, we provide access to the
transforms in convertMessage via the transforms() method on the 1st
param to convertMessage(), which provides the ROSTransforms object to
the converter.

This change provides classes to encapsulate and offer a richer object
than just a simple collection. To maintain backwards compatibility it
uses Proxy objects.